#477 Applying feedback: second round
