+++ examples/recording_service/pluggable_storage/c/FileStorageReader.c
@@ -23,24 +23,10 @@
 #include "routingservice/routingservice_infrastructure.h"
 
 #include "FileStorageReader.h"
+#include "FileStorageUtils.h"
 #include "HelloMsg.h"
 
 #define NANOSECS_PER_SEC 1000000000
-
-int RTI_fopen(FILE **file, const char *file_name, const char *mode)
-{
-    int error_code = 0;
-#ifdef RTI_WIN32
-    error_code = fopen_s(file, file_name, mode);
-#else
-    (*file) = fopen(file_name, mode);
-
-    if ((*file) == NULL) {
-        error_code = 1;
-    }
-#endif
-    return error_code;
-}
 
 #ifdef _WIN32
     #define SCNi64 "I64i"
